Generated by Spark: Created a new Contact and saved successfully. Good job, Thanks!
I, however, see that the region is a text field. It should be a drop-down of already-maintained Regions in the application, in the settings. Regions should be the continents of the world.
Also, add a country drop-down of all the countries in the world.  The region should be selected before the country, and the system should only list countries in the selected region. E.g., if Africa is selected, only African countries should come up for selection in the country field.